requirements.txt
asgiref==3.3.1
backcall==0.2.0
beautifulsoup4==4.9.3
colorama==0.4.4
decorator==4.4.2
Django==3.1.7
django-appconf==1.0.4
django-bootstrap-v5==1.0.0
django-extensions==3.1.1
django-imagekit==4.0.2
importlib-metadata==2.1.1
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pickleshare==0.7.5
pilkit==2.0
Pillow==8.1.2
prompt-toolkit==3.0.18
Pygments==2.8.1
pytz==2021.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
zipp==3.4.1
